+++ src/movement/index.js
@@ -1,0 +1,3 @@
+export * from './smooth-locomotion.component';
+export * from './smooth-turn.component';
+export * from './snap-turn.component';