+++ src/nav-mesh/index.js
@@ -1,0 +1,3 @@
+export * from './nav-mesh-strategy.component';
+export * from './nav-mesh.component';
+export * from './nav-mesh.system';