+++ src/nav-mesh/nav-mesh.component.js
@@ -1,0 +1,10 @@
+AFRAME.registerComponent('nav-mesh', {
+    schema: {},
+    init: function() {
+        this.system.registerNavMesh(this.el);
+    },
+
+    remove: function() {
+        this.system.unregisterNavMesh(this.el);
+    }
+});