+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  * Input mode property for switching between analog and binary input handling (`smooth-turn` and `smooth-locomotion`)
  * New RC car example
  * Vignette primitive `al-vignette` and corresponding example
+ * Initial nav-mesh support (`nav-mesh-strategy` component)
 
 # 0.1.0 (2022-05-03)
  * Smooth locomotion component `smooth-locomotion`